+++ posts/delete.php
@@ -3,35 +3,40 @@
 
 <?php
 
-    if(isset($_GET['del_id'])) {
-        $id = $_GET['del_id'];
+if (isset($_GET['del_id'])) {
+    $id = filter_input(INPUT_GET, 'del_id', FILTER_VALIDATE_INT);
 
-        $select = $conn->query("SELECT * FROM posts WHERE id = '$id'");
-        $select->execute();
-        $posts = $select->fetch(PDO::FETCH_OBJ);
-
-        if($_SESSION['user_id'] !== $posts->user_id) {
-             header('location: http://localhost:31337/project/index.php');
-            //echo "you?";
-        }
-        else {
-
-            $img = $posts->img;
-            if(!empty($img)) { // if images exist
-
-            unlink('C:/xampp/htdocs/project/posts/images' . $img );
-            }
-
-            $delete = $conn->prepare("DELETE FROM posts WHERE id = :id");
-            $delete->execute([
-                ':id' => $id
-            ]);
-
-        }
-        echo "<div class='alert alert-danger text-center role='alert'> Successfully deleted. </div>";
-        // header('location: http://localhost:31337/project/index.php');
-    } else {
-        header('location: http://localhost:31337/project/404.php');
+    // sanitize
+    if (!is_numeric($id)) {
+        header("Location: http://localhost:31337/project/404.php");
+        exit;
     }
+
+    $select = $conn->prepare("SELECT * FROM posts WHERE id = :id");
+    $select->bindParam(':id', $id);
+    $select->execute();
+    $posts = $select->fetch(PDO::FETCH_OBJ);
+
+    
+    if ($_SESSION['user_id'] !== $posts->user_id) {
+        header('location: http://localhost:31337/project/404.php');
+    } else {
+
+        $img = $posts->img;
+        if (!empty($img)) { // if images exist
+
+            unlink('C:/xampp/htdocs/project/posts/images' . $img);
+        }
+
+        $delete = $conn->prepare("DELETE FROM posts WHERE id = :id");
+        $delete->execute([
+            ':id' => $id
+        ]);
+    }
+    echo "<div class='alert alert-danger text-center role='alert'> Successfully deleted. </div>";
+    // header('location: http://localhost:31337/project/index.php');
+} else {
+    header('location: http://localhost:31337/project/404.php');
+}
 
 ?>
